asm/requirements.txt
click==8.1.3
Flask==2.0.3
flask-injector==0.13.0
connexion[swagger-ui]==2.13.1
dynaconf==3.1.8
Flask-Login
Flask-SQLAlchemy
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.2
python-dotenv==1.0.0
SQLAlchemy==1.4.41
Werkzeug==2.2.3
WTForms==3.0.1
gunicorn==20.1.0
validators==0.20.0
email_validator==1.3.1
grpcio==1.51.3
protobuf==4.22.0
